fix: public api for entities
