+++ src/entities/accsses/index.ts
@@ -1,0 +1,2 @@
+export type { Access } from './Access'
+export { getAllAccesses } from './api'
