+++ src/entities/role/index.ts
@@ -1,1 +1,2 @@
 export type { Role } from './Role'
+export { getAllRoles } from './api'
