+++ src/entities/user/api/index.ts
@@ -1,3 +1,4 @@
-export { getToken } from "./getToken";
-export type { Token } from "./getToken";
-export { getUserInfo } from "./getUserInfo";
+export { getToken } from './getToken'
+export type { Token } from './getToken'
+export { getUserInfo } from './getUserInfo'
+export { getAllUsers } from './getAllUsers'
