+++ src/entities/user/index.ts
@@ -1,2 +1,2 @@
-export { useUserStore, type UserInfo } from "./model"
-export { type Token, getToken, getUserInfo } from "./api"
+export { useUserStore, type UserInfo } from './model'
+export { type Token, getToken, getUserInfo, getAllUsers } from './api'
