+++ src/widgets/listLicenses/model/index.ts
@@ -1,2 +1,1 @@
-export { type Column } from './columnType'
 export { exportTable } from './tableExport'
